# live_openapi.json
{"openapi":"3.1.0","info":{"title":"ViralEdge M3 Engine","description":"Professional AI-powered viral content generator","version":"3.0"},"servers":[{"url":"/api/py"}],"paths":{"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/m3/analyze":{"get":{"summary":"M3 Analyze","operationId":"m3_analyze_m3_analyze_get","parameters":[{"name":"url","in":"query","required":true,"schema":{"type":"string","description":"Video URL","title":"Url"},"description":"Video URL"},{"name":"tier","in":"query","required":false,"schema":{"type":"string","description":"User Tier","default":"Free","title":"Tier"},"description":"User Tier"},{"name":"platform","in":"query","required":false,"schema":{"type":"string","description":"Platform (youtube/tiktok)","default":"youtube","title":"Platform"},"description":"Platform (youtube/tiktok)"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","description":"Comment limit","default":100,"title":"Limit"},"description":"Comment limit"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/m3/generate-script":{"post":{"summary":"Generate Script Endpoint","operationId":"generate_script_endpoint_m3_generate_script_post","parameters":[{"name":"title","in":"query","required":true,"schema":{"type":"string","description":"Video Title","title":"Title"},"description":"Video Title"},{"name":"duration","in":"query","required":false,"schema":{"type":"string","description":"Duration","default":"60s","title":"Duration"},"description":"Duration"},{"name":"tone","in":"query","required":false,"schema":{"type":"string","description":"Tone","default":"Energetic","title":"Tone"},"description":"Tone"},{"name":"notes","in":"query","required":false,"schema":{"type":"string","description":"Additional Notes","default":"","title":"Notes"},"description":"Additional Notes"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}